разбила код на модули
